fetch all users using useEffect
